refactor: move QuestBoardContent to commonMain
Splits QuestBoardContent out of QuestScreen.kt into commonMain (its
CreateQuestForm/QuestCard component dependencies were already moved in
the previous commit), enabling PreviewScreenshotGeneratorTest on the
JVM target.
